requirements.txt
streamlit==1.41.1
google-generativeai==0.8.4
pyyaml==6.0.2
python-dotenv>=1.0.0
protobuf>=4.25.1 